# docs/features/movies-on-demand/API_CONTRACT.yml
openapi: 3.1.0
info:
    title: Movies on Demand API
    description: >-
        API contract document defining the interface between frontend and backend.
        This serves as the single source of truth for coupling.
        Ephemeral movie streaming service using TMDB for discovery, NZBGeek for Usenet
        release search, and Cloudflare for on-demand acquisition and streaming.
        Movies are fetched from Usenet on-demand and automatically deleted after viewing.
    version: 1.0.0
    contact:
        name: API Support
servers:
    - url: http://localhost:8787/api
      description: Development Server
    - url: https://bahasadri.com/api
      description: Production Server
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        ErrorResponse:
            type: object
            required:
                - error
            properties:
                error:
                    type: string
                    description: A descriptive error message.
                    example: "Invalid input provided."
                code:
                    type: string
                    description: Machine-readable error code.
                    enum:
                        [
                            INVALID_INPUT,
                            NOT_FOUND,
                            INTERNAL_ERROR,
                            TMDB_ERROR,
                            NZBGEEK_ERROR,
                            USENET_ERROR,
                            STORAGE_ERROR,
                            JOB_ERROR,
                        ]
                    example: INVALID_INPUT
...
        StreamResponse:
            type: object
            required:
                - stream_url
                - content_type
            properties:
                stream_url:
                    type: string
                    description: URL to stream the movie
                    example: "https://r2.example.com/movies/job_abc123/movie.mp4"
                content_type:
                    type: string
                    description: MIME type of the video
                    example: "video/mp4"
                file_size:
                    type: integer
                    nullable: true
                    description: File size in bytes
                    example: 1572864000
...
    responses:
        BadRequest:
            description: Invalid input provided (400).
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    examples:
                        invalid_input:
                            value:
                                error: "Query parameter 'query' is required."
                                code: "INVALID_INPUT"
        NotFound:
            description: Resource not found (404).
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    examples:
                        not_found:
                            value:
                                error: "Movie with ID 123 not found."
                                code: "NOT_FOUND"
        InternalError:
            description: Server error (500).
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    examples:
                        server_error:
                            value:
                                error: "Unexpected database error."
                                code: "INTERNAL_ERROR"
...
        StorageError:
            description: Storage operation error (502).
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    examples:
                        storage_error:
                            value:
                                error: "Failed to store movie file."
                                code: "STORAGE_ERROR"
...
paths:
...
    /movies-on-demand/movies/{id}/stream:
        parameters:
            - name: id
              in: path
              required: true
              description: TMDB movie ID
              schema:
                  type: string
                  example: "550"
        get:
            tags:
                - Streaming
            operationId: getMovieStream
            summary: Get movie stream URL
            description: Retrieves the streaming URL for a movie that is ready. Returns a signed URL or direct stream URL from R2.
            responses:
                "200":
                    description: Movie stream information
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StreamResponse"
                "404":
                    $ref: "#/components/responses/NotFound"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "500":
                    $ref: "#/components/responses/InternalError"
                "502":
                    $ref: "#/components/responses/StorageError"
...
    /movies-on-demand/internal/progress:
        post:
            tags:
                - Internal
            operationId: reportJobProgress
            summary: Report job progress (Internal)
            description: >-
                Internal endpoint for containers to report job progress updates.
                Protected by Cloudflare Zero Trust service tokens.
                Not intended for frontend use - frontend should poll /jobs/{jobId} instead.
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - job_id
                                - status
                            properties:
                                job_id:
                                    type: string
                                    description: Job identifier
                                    example: "job_abc123"
                                status:
                                    type: string
                                    enum:
                                        [
                                            queued,
                                            downloading,
                                            uploading,
                                            ready,
                                            error,
                                        ]
                                    description: Current job status
                                    example: "downloading"
                                progress:
                                    type: number
                                    format: float
                                    nullable: true
                                    description: Progress percentage (0-100) for downloading or uploading
                                    example: 45.5
                                error_message:
                                    type: string
                                    nullable: true
                                    description: Error message if status is error
                                    example: "Download failed: Connection timeout"
            responses:
                "200":
                    description: Progress update accepted
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "500":
                    $ref: "#/components/responses/InternalError"